kafka-producers/requirements.txt
kafka-python
python-dotenv
fastapi[all]
pydantic
Faker